refactor: split docs screenshot tooling and simplify validation
